+++ tipg/factory.py
@@ -439,10 +439,14 @@ class OGCFeaturesFactory(EndpointsFactory):
                 },
             },
         )
-        def collections(
+        def collections(  # noqa: C901
             request: Request,
             bbox_filter: Annotated[Optional[List[float]], Depends(bbox_query)],
             datetime_filter: Annotated[Optional[List[str]], Depends(datetime_query)],
+            type_filter: Annotated[
+                Optional[Literal["Function", "Table"]],
+                Query(alias="type", description="Filter based on Collection type."),
+            ] = None,
             limit: Annotated[
                 Optional[int],
                 Query(
@@ -471,6 +475,14 @@ class OGCFeaturesFactory(EndpointsFactory):
 
             limit = limit or 0
             offset = offset or 0
+
+            # type filter
+            if type_filter is not None:
+                collections_list = [
+                    collection
+                    for collection in collections_list
+                    if collection.type == type_filter
+                ]
 
             # bbox filter
             if bbox_filter is not None: