+++ CHANGES.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 Note: Minor version `0.X.0` update might break the API, It's recommended to pin `tipg` to minor version: `tipg>=0.1,<0.2`
 
+## unreleased
+
+### Added
+
+* `type` query parameter to filter collections based on their type (`Function` or `Table`)
+
 ## [0.2.0] - 2023-06-22
 
 ### Changed